+++ examples/custom_dictionaries.rs
@@ -1,11 +1,12 @@
 // NOTE: Every `.unwrap()` seen here is SAFE.
 
-use names::GeneratorBuilder;
+use names::{GeneratorBuilder, Generator};
+use rand::rngs::ThreadRng;
 
 fn main() {
-    let adjectives = vec!["imaginary"];
-    let nouns = vec!["roll"];
-    let mut generator = GeneratorBuilder::default()
+    let adjectives: Vec<String> = vec!["imaginary".into()];
+    let nouns: Vec<String> = vec!["roll".into()];
+    let mut generator: Generator<ThreadRng> = GeneratorBuilder::default()
         .adjectives(adjectives)
         .nouns(nouns)
         .build()