+++ examples/default.rs
@@ -1,8 +1,8 @@
 // NOTE: Every `.unwrap()` seen here is SAFE.
 
-use names::Generator;
+use names::{Generator, ThreadRng};
 
 fn main() {
-    let mut generator = Generator::default();
+    let mut generator = Generator::<ThreadRng>::default();
     println!("Your project is: {}", generator.next().unwrap());
 }