+++ examples/serde.rs
@@ -1,7 +1,9 @@
 // NOTE: Every `.unwrap()` seen here is SAFE, just make sure that `adjectives` and `nouns` are not empty.
 
+use rand::rngs::ThreadRng;
+
 fn main() {
-    let mut generated = serde_json::from_str::<names::Generator>(r#"{
+    let mut generated = serde_json::from_str::<names::Generator<ThreadRng>>(r#"{
         "casing": "CamelCase",
         "naming": {
             "ZeroPaddedNumbered": [2, "_"]