+++ examples/trail_4_digit.rs
@@ -1,9 +1,10 @@
 // NOTE: Every `.unwrap()` seen here is SAFE.
 
-use names::{GeneratorBuilder, Name, NumberSeperator};
+use names::{GeneratorBuilder, Name, NumberSeperator, Generator};
+use rand::rngs::ThreadRng;
 
 fn main() {
-    let mut generator = GeneratorBuilder::default()
+    let mut generator: Generator<ThreadRng> = GeneratorBuilder::default()
         .naming(Name::ZeroPaddedNumbered(4, NumberSeperator::Dash))
         .build()
         .unwrap();